Backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.3
sqlalchemy==2.0.41
psycopg[binary]==3.2.9
python-dotenv==1.1.0
pandas==2.2.3
openpyxl==3.1.5
langchain==0.3.25
langchain-groq==0.3.2
langgraph==0.4.7
pydantic==2.11.4
